requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
idna==3.10
pillow==11.2.1
PyJWT==2.10.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyserial==3.5
qrcode==8.1
requests==2.32.3
urllib3==2.4.0
websocket-client==1.8.0
